Make extension attributes private in codegen rules (#123)
These attributes should be controlled by codegen authors only, but
because they are used in computing outputs they needed to be public.
Bazel has been updated to allow private attributes as argument to the
outputs method.

Drop split/join logic for out_extensions. This was only in place since
attr.string_list doesn't have a `values`argument to allow us to limit to
the only correct value. Since it is private that extra protection isn't
necessary.