Add files which may be relative imports as srcs (#126)
The URI inside a summary for a file outside of lib will use the `file`
scheme. When producing a Source for a file outside of lib, codegen with
summaries will use an `asset` scheme URI. Relative imports from this
source will therefore also use an `asset` scheme, which means they will
not match the URI contained in summary files. The effect is that the
`imports` iterable on `LibraryElement` would omit these sources.

Long term we may want to produce summaries with `asset` scheme URIs. In
the short term we can ensure these files are readable with this scheme
by including them as src files.