Update `make_dart_context` to take `data` as a list of targets rather than (#174)
a list of files.

This change updates `make_dart_context` such that `data` is now a list of
targets rather than a list of files.  References to those targets are
maintained and exposed in the `transitive_data` struct field so that callers
may in turn access the providers of those targets.

Most of the time, however, callers of `dart_ctx.transitive_data` still just
want the files (to add it to their runfiles, for example).  So for both
caller convenience and efficiency, `make_dart_context` still maintains the
depset as well.  It now exposes the `transitive_data` struct field as follows:

-- Before

    struct(
        ...
        transitive_data = <depset>,
    )

-- After

    struct(
        ...
        transitive_data = struct(
            targets = {<label_name>: <target>},
            files: <depset>,
        ),
    )

For an attempt at API consistency, this CL updated any property of the
`dart_ctx` struct whose name matches "transitive_*" to follow the same
pattern, but to omit the dict or depset if it doesn't apply.  So we have:

    struct(
        ...
        transitive_srcs = struct(
            targets = {<label_name>: <target>},
            files = <depset>,
        ),
        transitive_dart_srcs = struct(
            files = <depset>,
        ),
        transitive_data = struct(
            targets = {<label_name>: <target>},
            files = <depset>,
        ),
        transitive_deps = struct(
            targets = {<label_name>: <target>},
        ),
        transitive_archives = struct(
            files = <depset>,
        ),
    )

You'll notice above that `transitive_srcs`, which used to be a depset, now
contains both the dict and the depset.  That's because in at least one
place, we were taking the transitive sources from a dart context and passing
them as data to `dart_vm.binary_action` (which in turn passed the data
through to `make_dart_context`).  Thus, `transitive_srcs` needs to expose
the targets as well, which in turn means that the `srcs` argument to
`make_dart_context` needs to accept a list of targets.

The tricky part about `srcs`, however, is that we can't force the attribute
to only contain targets; sometimes, generated outputs (which have no targets)
exist in this list.  To deal with this, this CL introduces a new parameter to
`make_dart_context`: `generated_srcs`, which is a list of File.  In the case
`generated_srcs`, they simply won't contribute to the `transitive_srcs.targets`
return value.